AI/requirements.txt
# === CORE API ===
fastapi
uvicorn[standard]
pydantic

# === DATA / UTIL ===
python-dotenv
python-multipart
httpx
veryfi
huggingface-hub

# === AI / NLP ===
torch
transformers
accelerate
bitsandbytes
underthesea
sentence-transformers

# === EVALUATION / PLOTTING ===
matplotlib
pandas
tabulate

# === EXPORT ===
python-docx
